requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
click==7.1.2
configparser==5.0.1
dataclasses==0.7
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
paho-mqtt==1.5.1
pathspec==0.8.0
pkg-resources==0.0.0
playsound==1.2.2
pylint==2.6.0
regex==2020.10.28
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
